An initial look at the web objects, the JAXB calls, and the way to make transforms out of objects.
Still a lot of work to do before viability, including storing and referring to types in the system, building the transform lists, and reading the type configurations from the config directory structure.

But this should serve as a good introduction to the concepts and what we CAN do given the system as it is.